move BuildsQueries under Query contracts
